All movies logic placed in movies.mjs
